refactor: split inspector panel
